+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - 修复结合 `vue-i18n` 使用时会提示不能覆盖 `$t` 方法的问题
 - 新增 Loading 自定义加载文案的功能，并优化了视觉表现
 - 修复 Input blur 事件的参数不是 event 对象的问题
+- `window.ELEMENT.locale.use` 改成 `window.ELEMENT.locale`
 
 ### 1.0.0-rc.9